Implemented IMAP METADATA (RFC5464) support for both the socket and cclient based drivers.
The code should support both the final RFC version of the extension as well as a predecessor of the RFC called ANNOTATEMORE. Currently Cyrus Imapd implements ANNOTATEMORE. There is also a dovecot plugin. The IMAP METADATA extension is central to Kolab support.

Added a mock driver with the primary objective of supporting unit tests for the Kolab drivers.