Add Horde_Imap_Client_Base::parseCacheId().
This makes all cache ID generation internal to Horde_Imap_Client, which
is desirable.

Additionally, for CONDSTORE hosts, we should use the last modseq seen
with respect to the *browser*, not the server.  This ensures that all
changes are caught when the browser does a request - if a browser misses
a request, for example, it may not catch the changes.