Rewrite IMAP command generator.
An unfortunate IMAP implementation detail is the various ways strings
are handled.  Depending on the command, and different portions of the
command, different quoting styles are needed.  And to make things worse,
literals are always needed in certain cases (if CR or LF is present in
the string; if 8-bit characters are present).

Refactor the _sendLine() function to accept array input, with the
various elements of the array identified as the various IMAP string
types.  Then, in a central location, figure out how to quote/send the
command to the server.