[mms] Add support for IMAP MOVE command.
Although not yet an RFC standard, at last count over 11 servers
support (including Dovecot and Cyrus).  A big performance win wrapping
three distinct commands (and their corresponding cache writes) into a
single, atomic command.

Also, the main reason that I discovered our IMAP parsing was lacking in
some ways.  This command may cause things like COPYUIDs, FETCHs, and
EXPUNGE untagged responses to all appear in the same command. Thus, we
need to handle each of these untagged responses as they come up, rather
than relying on a specific command call to deal with these commands
after the parent command is done.

It also highlights the vagaries of QRESYNC, which require the client to
continue to process message sequence numbers as there is no guarantee
that VANISHED responses will be sent (theoretically, VANISHED and
EXPUNGE responses could be sent in the same untagged sequence - it is
critical for a client to be able to handle this sort of output).

Now that our underlying handling has been fixed, the MOVE command is
trivial to implement, as it should be, since it is nothing more than a
pseudo-macro of three existing IMAP commands (COPY - STORE - EXPUNGE).